sửa và cập nhật lại các lỗi và dư thừa
